requirements.txt
dj-database-url==0.5.0
Django==2.1.2
django-admin==1.3.2
django-excel-response2==2.0.8
django-filter==2.0.0
django-six==1.0.4
djangorestframework==3.8.2
gunicorn==19.9.0
Markdown==2.6.11
psycopg2==2.7.5
Pygments==2.2.0
pytz==2018.5
whitenoise==4.1
xlwt==1.3.0
django-cors-headers==2.4.0
